Added change handler to search bar
